requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
bilibili-api==9.0.1
blinker==1.4
Brotli==1.0.9
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
h11==0.12.0
h2==4.0.0
hpack==4.0.0
httptools==0.3.0
Hypercorn==0.11.2
hyperframe==6.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.4
MarkupSafe==2.0.1
multidict==5.2.0
nest-asyncio==1.5.1
priority==2.0.0
pyasn1==0.4.8
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
sanic==21.9.3
sanic-routing==0.7.2
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
websockets==10.0
Werkzeug==2.0.1
wsproto==1.0.0
yarl==1.6.3
moviepy~=1.0.3